BoardConfigQcom: Fix setting UM4.14 specific flags
Calling is-board-platform-in-list from this makefile was not returning the desirable output, thus resulting in a failure to meet the conditions for the UM4.14 family platforms.

Due to the fact that this makefile is included in a very early stage in the build process, the QCOM macros are not defined. Simply use standard macros like in rest of the makefile.

Change-Id: Iad889882f94167f1bc1761ddd549dac9dcdee3ba